requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
gradio>=4.0.0


